cleanup unneeded fxlib.h includes, move to atan2f, extend near-plane culling limit to avoid some artifacting

+++ worldobj.c
@@ -21,7 +21,7 @@ fixed angle_horizontal_plane(vec3f u, vec3f v) {
 	fixed dot = mulff(u.x, v.x) + mulff(u.z, v.z); // dot product
 	fixed det = mulff(u.x, v.z) - mulff(u.z, v.x); // determinant
 	// just do it with floats and atan2, this only runs a few times in a tick
-	fixed angle = float2f((float)atan2(f2float(det), f2float(dot)));
+	fixed angle = float2f(atan2f(f2float(det), f2float(dot)));
 	return angle;
 }
